Avaces en las apuestas

+++ proyecto_rog_clicker/src/proyecto_rog_clicker/ElementoJuego.java
@@ -9,7 +9,7 @@ public class ElementoJuego {
 		this.nombre = nombre;
 	}
 	
-	public void efecto(int boost_clicks, int boost_prod, int boost_cant, long produccion, long clicks, long cant) {
+	public void efecto(long boost_clicks, long boost_prod, long boost_cant, long produccion, long clicks, long cant) {
 		
 	}
